backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
anthropic==0.7.7
httpx==0.25.1
pydantic==2.5.0
python-dotenv==1.0.0
websockets==12.0
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6 